WIS-151-delete-wishlist-product-in-delete-wishlist-and-delete-product fixed the deletion of wishlist
Added a JPQL statement to WishlistRepository to delete the wishlist and wrote a method in WishlistProductRepository to delete WishlistProduct

+++ src/main/java/com/caldev/wishlister/services/ProductService.java
@@ -134,7 +134,17 @@ public class ProductService {
 
     @Transactional
     public void deleteProduct(Long requestedProductId) {
-        wishlistProductService.deleteWishlistProductByProductId(requestedProductId); // cascade delete()
+
+        Optional<Product> productToDelete = findProductById(requestedProductId);
+
+        if (productToDelete.isEmpty()) {
+            throw new ProductNotFoundException("Product not found");
+        }
+
+        for (WishlistProduct wishlistProduct : productToDelete.get().getWishlistProducts()) {
+            wishlistProductService.deleteWishlistProduct(wishlistProduct.getWishlistProductId());
+        }
+
         productRepository.deleteById(requestedProductId);
     }
